Scripts for wordeval on hpc
3 Scripts:
Timelm_word_eval makes a csv file containing the probability of each word being assigned to each class by each model. this is meant to be ran by submit_word.sh
submit_word.sh is a shell script to run Timelm_word_eval.py on the hpc as there were trouble downloading the models created on the hpc.
Last script cleans up the csv file created by Timelm_word_eval.py such that the probabilities are saved as a string with spaces in between with Negative first, neutral second and positive third. The numbers are also rounded to 4 decimals