Word_eval_cleaner fixes
the last model put "LABEL:0" or  "LABEL:1" or "LABEL:2"  into the string in the dataframe, since numbers persisted this made the last model seem like a above 100% probability of being negative. It is fixed by removing the part where this "LABEL:#" was.
A few other comments have been added for readability